refactor: remove unnecessary `.toInt()` conversions in Uri and VideoPlayerState classes
Simplified code by removing redundant `.toInt()` calls for `rowBytes` and `path`, ensuring consistency and clarity across macOS and Linux platform implementations.

+++ mediaplayer/src/jvmMain/kotlin/io/github/kdroidfilter/composemediaplayer/linux/LinuxVideoPlayerState.kt
@@ -505,7 +505,7 @@ class LinuxVideoPlayerState : VideoPlayerState {
 
                     // Native-to-native copy: frame buffer -> Skia bitmap pixels
                     srcBuf.rewind()
-                    val destRowBytes = pixmap.rowBytes.toInt()
+                    val destRowBytes = pixmap.rowBytes
                     val destSizeBytes = destRowBytes.toLong() * height.toLong()
                     val destBuf =
                         LinuxNativeBridge.nWrapPointer(pixelsAddr, destSizeBytes)
